new variant training prelayernorm #2 keep training

+++ parlai/gcp/gcs_service.py
@@ -6,6 +6,7 @@ import os
 from os import environ, listdir
 from os.path import isfile, join
 from io import BytesIO, StringIO
+from parlai.utils.misc import warn_once
 
 from google.cloud import storage
 
@@ -83,7 +84,8 @@ class GCP_Service(object):
                 logger.info(f'{dest_file} downloaded from bucket.')
             except:
                 print('raise exception')
-                os.remove(dest_file)
+                if os.path.isfile(dest_file):
+                    os.remove(dest_file)
                 continue
 
     def download(self, bucket_filename:str, local_dir:str) -> str:
@@ -97,8 +99,9 @@ class GCP_Service(object):
             logger.info(f'{file_name} downloaded from bucket.')
             return dest_file
         except:
-            os.remove(dest_file)
-            logger.warn(f'{bucket_filename} not found in {self.bucket_name}')
+            if os.path.isfile(dest_file):
+                os.remove(dest_file)
+            warn_once(f'{bucket_filename} not found in {self.bucket_name}')
 
     def delete(self, bucket_filename:str):
         self.bucket.delete_blob(bucket_filename)
